airflow/dags/requirements.txt
bs4
boto3
apache-airflow
pyarrow
requests
urllib
snowflake-connector-python
python-dotenv
pandas